requirements.txt
functions-framework==3.*
pytz~=2023.3
yahoo_fin~=0.8.9
requests~=2.28.2
pandas~=2.0.2
pandas_gbq
pandas.io
gcsfs
google-cloud-bigquery
google-cloud-logging
